+++ lua/rustaceanvim/config/internal.lua
@@ -42,13 +42,20 @@ local RustaceanDefaultConfig = {
     ---@class RustaceanHoverActionsConfig
     hover_actions = {
 
-      --- whether to replace Neovim's built-in `vim.lsp.buf.hover`
+      --- whether to replace Neovim's built-in `vim.lsp.buf.hover`.
+      --- default: true
       ---@type boolean
       replace_builtin_hover = true,
+    },
 
-      -- the border that is used for the hover window
+    --- options same as lsp hover
+    ---@see vim.lsp.util.open_floating_preview
+    ---@type table Options applied to floating windows.
+    float_win_config = {
+
+      -- the border that is used for floating windows
       ---@see vim.api.nvim_open_win()
-      ---@type string[][]
+      ---@type string[][] | string
       border = {
         { '╭', 'FloatBorder' },
         { '─', 'FloatBorder' },
@@ -58,17 +65,17 @@ local RustaceanDefaultConfig = {
         { '─', 'FloatBorder' },
         { '╰', 'FloatBorder' },
         { '│', 'FloatBorder' },
-      },
+      }, -- maybe: 'double', 'rounded', 'shadow', 'single',
 
-      --- maximal width of the hover window. Nil means no max.
+      --- maximal width of floating windows. Nil means no max.
       ---@type integer | nil
       max_width = nil,
 
-      --- maximal height of the hover window. Nil means no max.
+      --- maximal height of floating windows. Nil means no max.
       ---@type integer | nil
       max_height = nil,
 
-      --- whether the hover action window gets automatically focused
+      --- whether the window gets automatically focused
       --- default: false
       ---@type boolean
       auto_focus = false,
@@ -301,7 +308,6 @@ local RustaceanDefaultConfig = {
   },
   was_g_rustaceanvim_sourced = vim.g.rustaceanvim ~= nil,
 }
-
 local rustaceanvim = vim.g.rustaceanvim or {}
 local opts = type(rustaceanvim) == 'function' and rustaceanvim() or rustaceanvim
 if opts.tools and opts.tools.executor and type(opts.tools.executor) == 'string' then