+++ lua/rustaceanvim/hover_actions.lua
@@ -78,17 +78,20 @@ function M.handler(_, result, ctx)
     return
   end
 
+  -- NOTE: This is for backward compatibility
+  local win_opt = vim.tbl_deep_extend('force', config.tools.float_win_config, config.tools.hover_actions)
+
   local bufnr, winnr = lsp_util.open_floating_preview(
     markdown_lines,
     'markdown',
-    vim.tbl_extend('keep', config.tools.hover_actions, {
+    vim.tbl_extend('keep', win_opt, {
       focusable = true,
       focus_id = 'rust-analyzer-hover-actions',
       close_events = { 'CursorMoved', 'BufHidden', 'InsertCharPre' },
     })
   )
 
-  if config.tools.hover_actions.auto_focus then
+  if win_opt.auto_focus then
     vim.api.nvim_set_current_win(winnr)
   end