+++ CHANGELOG.md
@@ -6,6 +6,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Config: `tools.float_win_config` for all floating Windows
+  created by this plugin.
+  Moved `border`, `max_width`, `max_height`, `auto_focus`
+  from `hover_actions` to `float_win_config`.
+  The `hover_actions` window options are still applied
+  if they exist, so as not to break compatibility.
+
 ## [3.12.2] - 2024-01-07
 
 ### Fixed